[flang][driver] Support -fdebug-pass and -mdebug-pass in flang
The options are used in clang to list the codegen passes that are run by the
legacy pass manager. This beings flang's behavior in line with clang. Enabling
these requires setting an LLVM command line option. As a consequence, the wart
in the implementation is that a fake command line is generated and parsed if
-fdebug-pass or -mdebug-pass is provided to flang. This also mirrors the
implementation in clang,