[flang][driver] Add options -fdefault-integer-4 and -fdefault-real-4
These options are supported by the Intel compiler. There are similar options in
the IBM compiler as well. Each of the options can be provided more than once.
In that case, the last occurence of -fdefault-real-* (respectively
-fdefault-integer-*) is used if both -fdefault-real-4 and -fdefault-real-8 are
specified. The fdefault.f90 test file was extensively modified. The handling of
the option by the driver was removed since testing the handling of the options
by -fc1 is sufficient.

Fixes #160252